+++ src/HDLparser/parser/index.js
@@ -1,0 +1,2 @@
+const parser = require('./parser');
+