+++ src/HDLtool/doc/html.js
@@ -7,6 +7,7 @@ const { getIconConfig } = require('../../HDLfilesys/icons');
 
 const opeParam = require('../../param');
 const HDLPath = require('../../HDLfilesys/operation/path');
+const { RenderType, Global } = require('./common');
 
 const _cache = {
     css : null
@@ -22,14 +23,61 @@ function makeFinalHTML(body, style) {
     <title>Document</title>
 </head>
 <body>
-    <div id="write">
-        ${body}
+    <div id="wrapper">
+        <div id="write">
+            ${body}
+        </div>
     </div>
 </body>
 <style>
     ${style}
 </style>
 </html>`;
+}
+
+
+function makeExportHTML(cssHref, body) {
+    return `<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Document</title>
+    <link rel="stylesheet" type="text/css" href="${cssHref}"></link>
+</head>
+<body>
+    <div id="wrapper">
+        <div id="write">
+            ${body}
+        </div>
+    </div>
+</body>
+</html>`;
+}
+
+function makeCommonElement(renderResult) {
+    return renderResult + '<br>\n';
+}
+
+function makeSVGElement(renderResult, caption) {
+    let mainHtml;
+    if (caption) {
+        mainHtml = '<div align=center>' + renderResult + `<p class="ImgCaption">${caption}</p>` + '</div>';
+    } else {
+        mainHtml = '<div align=center>' + renderResult + '</div>';
+    }
+    return '<br>' + mainHtml + '<br><br>\n';
+}
+
+function makeSVGElementByLink(link, caption) {
+    let mainHtml;
+    if (caption) {
+        mainHtml = `<div align=center><img src="${link}"></img><p class="ImgCaption">${caption}</p></div>`;
+    } else {
+        mainHtml = `<div align=center><img src="${link}"></img></div>`;
+    }
+    return '<br>' + mainHtml + '<br><br>\n';
 }
 
 function getDocCssString() {
@@ -47,27 +95,50 @@ function getDocCssString() {
 function makeWavedromRenderErrorHTML() {
     return `<div class="error-out">
     <p class="error">Error Render</p>
-</div>`;
+</div><br>`;
 }
 
-function showDocWebview() {
+
+/**
+ * @description make the html string of a finial display style
+ * @param {string} usage in whick module is used
+ * @returns {string}
+ */
+function makeShowHTML(usage) {
     const renderList = getCurrentRenderList();
+    if (renderList.length == 0) {
+        return '';
+    }
 
     // start to render the real html
     let body = '';
     for (const r of renderList) {
         const renderResult = r.render();
         if (renderResult) {
-            body += renderResult + '\n';
+            if (r.type == RenderType.Markdown) {
+                body += makeCommonElement(renderResult);
+            } else if (r.type == RenderType.Wavedrom) {
+                body += makeSVGElement(renderResult, r.desc);
+            }
         } else {
-            body += makeWavedromRenderErrorHTML() + '\n';
+            body += makeWavedromRenderErrorHTML();
         }
     }
 
     // add css
-    let style = getDocCssString();
-    const html = makeFinalHTML(body, style);
+    let cssString = getDocCssString();
+    if (usage == 'webview') {          // if invoked by webview, change background image
+        const webviewConfig = vscode.workspace.getConfiguration("TOOL.doc.webview");
+        const imageUrl = webviewConfig.backgroundImage;
+        cssString = cssString.replace("--backgroundImage", imageUrl);
+    } else if (usage == 'pdf') {      // if invoked by pdf, transform .vscode-light to #write
+        cssString = cssString.replace(/\.vscode-light/g, '#write');
+    }
+    const html = makeFinalHTML(body, cssString);
+    return html;
+}
 
+function showDocWebview() {
     const webview = vscode.window.createWebviewPanel(
         'TOOL.showDocWebview', 
         'document',
@@ -78,15 +149,77 @@ function showDocWebview() {
         }
     );
     webview.iconPath = getIconConfig('documentation');
-    webview.webview.html = html;
+    webview.webview.html = makeShowHTML("webview");
 }
 
 
 function exportCurrentFileDocAsHTML() {
+    if (vscode.window.activeColorTheme.kind != vscode.ColorThemeKind.Light) {
+        vscode.window.showErrorMessage('Please export html in a light theme!');
+        return;
+    }
 
+    const editor = vscode.window.activeTextEditor;
+    const currentFilePath = HDLPath.toSlash(editor.document.fileName);
+    const HDLFileName = HDLPath.basename(currentFilePath);
+    const renderList = getRenderList(currentFilePath);
+    if (renderList.length == 0) {
+        return;
+    }
+
+    const wsPath = opeParam.workspacePath;
+    const markdownFolderPath = HDLPath.join(wsPath, 'html');
+    if (!fs.existsSync(markdownFolderPath)) {
+        fs.mkdirSync(markdownFolderPath);
+    }
+    const currentRoot = HDLPath.join(markdownFolderPath, HDLFileName);
+    if (fs.existsSync(currentRoot)) {
+        HDLPath.deleteFolder(currentRoot);
+    }
+    fs.mkdirSync(currentRoot);
+    const figureFolder = HDLPath.join(currentRoot, 'figure');
+    fs.mkdirSync(figureFolder);
+
+    const cssFolder = HDLPath.join(currentRoot, 'css');
+    fs.mkdirSync(cssFolder);
+    const relateCssPath = './css/index.css';
+    const cssPath = HDLPath.join(cssFolder, 'index.css');
+    let cssString = getDocCssString();
+
+    // only support export in the ligth theme
+    cssString = cssString.replace(/\.vscode-light/g, '#write');
+    fs.writeFileSync(cssPath, cssString);
+
+    let body = '';
+    for (const r of renderList) {
+        const renderResult = r.render();
+        if (r.type == RenderType.Markdown) {
+            body += makeCommonElement(renderResult);
+        } else if (r.type == RenderType.Wavedrom) {
+            const svgName = 'wavedrom-' + Global.svgMakeTimes + '.svg';
+            const svgPath = HDLPath.join(figureFolder, svgName);
+            fs.writeFileSync(svgPath, renderResult);
+            const relatePath = HDLPath.join('./figure', svgName);
+            body += makeSVGElementByLink(relatePath, r.desc);
+        }
+    }
+
+    const html = makeExportHTML(relateCssPath, body);    
+    const htmlName = 'index.html';
+    const htmlPath = HDLPath.join(currentRoot, htmlName);
+    fs.writeFileSync(htmlPath, html);
+}
+
+
+function exportProjectDocAsHTML() {
+    vscode.window.showInformationMessage('this is exportProjectDocAsHTML');
 }
 
 
 module.exports = {
-    showDocWebview
+    showDocWebview,
+    exportCurrentFileDocAsHTML,
+    exportProjectDocAsHTML,
+    makeShowHTML,
+    makeSVGElementByLink
 };